cisco-nxos-provider: implement `vrf` context
* cisco-nxos-provider: DeletingUpdate implementation

The gnmiext client does not provide a way to delete objects from the
data tree. Until now we worked with static objects. This is now
different with, e.g., VRFs, which can be arbitrarily added and removed.

DeviceConf methods like Reset() can now return the DeletingUpdate
implementation of the Update interface. If a call to  gnmiext.Update
happens and the update objects are of this type then the gnmiext.Update
will issue a gNMI request with mode `delete` to remove the objects
contained in the data tree specified by the xpath.

* cisco-nxos: Add VRF context implementation

The minimal configuration options for a VRF on cisco-nxos is the
following:
```
vrf context CC-CLOUD01
  vni 101 l3
  rd 4269539332:101
  address-family ipv4 unicast
    route-target import 65148:1101
    route-target import 65148:1101 evpn
    route-target import 65148:4101
    route-target import 65148:4101 evpn
    route-target export 65148:1101
    route-target export 65148:1101 evpn
    route-target export 65148:4101
    route-target export 65148:4101 evpn
```

Attempts to configure a VRF on a device without the following features
already enabled will fail:
- feature vn-segment
- feature nv overlay
- nv overlay evpn
- feature bgp