cisco-nxos-provider: Activate IS-IS feature as config is applied (#35)
This patch adds another update call to automatically activate the IS-IS
feature when an IS-IS configuration item is applied. On the opposite
note, the feature will not be deactivated, once an IS-IS item is reset,
as there could potentially be other IS-IS processes running on the same
device that would be influenced.