cisco-nxos-provider: decouple ISIS from interface
We have decided to decouple the sections related to routing
configuration from the interface configuration itself. Hence, we
introduce changes to the pkg `isis` that allow a minimal configuration
of the ISIS properties for an interface.

These changes allow the equivalent to following CLI commands on Cisco:
```
interface eth1/1
ip router isis UNDERLAY
isis network-point-to-point
```

The current implementation assumes that the interfaces are properly
configured (they are "routed" interfaces).