Enable dynamic provider selection via environment variable in Tilt
This is set via an `PROVIDER` environment variable which is
automatically read when the tilt setup is started. The value should
match the provider name that will be provided as the `--provider`
argument to the manager.

The provider argument will be replaced in the kustomize rendered
manifests. This ensures that the override will be kept when Tilt reloads
the controller-manager deployment.