Remove finalizer for device credential secret (#95)
Previously the `Device` reconciler added a finalizer on the credentials
secret to it's deletion. This had a number of undesired side-effects.
1. This meant that the secret had to be immutable, as changing the
   secret reference on the device would leave previously referenced
   secrets with a finalizer, preventing their deletion.
2. This restriction and handling felt unlogical and inconsistent, as we
   didn't handle other secret references, such as the references used to
   refer to the tls credentials, in the same way.

With this PR we are removing the logic of adding finalizer logic on the
referenced credentials secret. If a referenced secret gets deleted,
while it's still being referred to by a `Device` resource, we let the
reconcilation fail and return an error. It will then be an
administrators responsibility to ensure that the referenced resource
exists, such that the `Device` (and child resources) can successfully be
reconciled. This is also inline with the overall consensus of handling
such situations within the wider ecosystem and what is done in other
projects.