Introduce operational condition on the Interface
This change introduces a new function on the `InterfaceProvider` to
allow retrieving the operational state (operSt on Cisco) of an Interface
resource and will set this as a status condition on the Kubernetes
resource. The top-level `Ready` condition is then computed based on the
`Configured` and `Operational` condition.

The retrieval of the operational state is also implemented for the Cisco
NX-OS provider.