Introduce `Aggregate` type for `Interface` resource
This change introduces a new `type` of `Interface` specified as
`Aggregate`. These interfaces represent aggregate ethernet interfaces
(also called `port-channel` on Cisco devices) that are used to build
`LinkAggregationGroup`s (LAG for short, also refered to as "Bundle
Interfaces").

This resource is only concerned with the configuration of the aggregate
(port-channel) but not with the orchestration of groups of interfaces
that are configured together to form a LAG.

As an aggregate references multiple `Interface`s of type `Physical`, the
respective members will be claimed by a `memberOf` field in the
`Interface` status. Additionally, a label is attached to refer to the
resource name of the `Aggregate` `Interface`.