Reconcile core resource when referenced provider config is updated
This changes introduces the capability for Resources used as provider
configuration to register themselves to trigger a reconciliation of the
core resource they are linked against. This is used on the
`ManagementAccess` resource, which will be reconciled, when a referenced
`ManagementAccessConfig` from the
`nx.cisco.networking.metal.ironcore.dev` api group gets updated.